Finalized dual weighted residual mesh adaptation sshock case
