+++ src/parameters/all_parameters.cpp
@@ -112,7 +112,7 @@ void AllParameters::declare_parameters (dealii::ParameterHandler &prm)
                       " burgers_rewienski_snapshot |"
                       " advection_periodicity | "
                       " flow_solver | "
-                      " dual_weighted_residual_convergence"),
+                      " dual_weighted_residual_mesh_adaptation"),
                       "The type of test we want to solve. "
                       "Choices are (only run control has been coded up for now)" 
                       " <run_control | " 
@@ -136,7 +136,7 @@ void AllParameters::declare_parameters (dealii::ParameterHandler &prm)
                       "  burgers_rewienski_snapshot |"
                       "  advection_periodicity | "
                       "  flow_solver | "
-                      "  dual_weighted_residual_convergence>.");
+                      "  dual_weighted_residual_mesh_adaptation>.");
 
     prm.declare_entry("pde_type", "advection",
                       dealii::Patterns::Selection(
@@ -225,7 +225,7 @@ void AllParameters::parse_parameters (dealii::ParameterHandler &prm)
     else if (test_string == "euler_naca0012")                       { test_type = euler_naca0012; }
     else if (test_string == "optimization_inverse_manufactured")    { test_type = optimization_inverse_manufactured; }
     else if (test_string == "flow_solver")                          { test_type = flow_solver; }
-    else if (test_string == "dual_weighted_residual_convergence")   { test_type = dual_weighted_residual_convergence; }
+    else if (test_string == "dual_weighted_residual_mesh_adaptation")   { test_type = dual_weighted_residual_mesh_adaptation; }
     
     const std::string pde_string = prm.get("pde_type");
     if (pde_string == "advection") {
