+++ src/testing/tests.cpp
@@ -30,7 +30,7 @@
 #include "burgers_rewienski_snapshot.h"
 #include "reduced_order.h"
 #include "flow_solver.h"
-#include "dual_weighted_residual_convergence.h"
+#include "dual_weighted_residual_mesh_adaptation.h"
 
 namespace PHiLiP {
 namespace Tests {
@@ -146,8 +146,8 @@ std::unique_ptr< TestsBase > TestsFactory<dim,nstate,MeshType>
         if constexpr (dim==2 && nstate==dim+2) return std::make_unique<EulerNACA0012<dim,nstate>>(parameters_input);
     } else if(test_type == Test_enum::flow_solver) {
         if constexpr (dim==3 && nstate==dim+2) return FlowSolverFactory<dim,nstate>::create_FlowSolver(parameters_input);
-    } else if(test_type == Test_enum::dual_weighted_residual_convergence) {
-        if constexpr (dim > 1)  return std::make_unique<DualWeightedResidualConvergence<dim, nstate>>(parameters_input);
+    } else if(test_type == Test_enum::dual_weighted_residual_mesh_adaptation) {
+        if constexpr (dim > 1)  return std::make_unique<DualWeightedResidualMeshAdaptation<dim, nstate>>(parameters_input);
     } else{
         std::cout << "Invalid test. You probably forgot to add it to the list of tests in tests.cpp" << std::endl;
         std::abort();
